#84 #83 notes linked with user

+++ be/src/modules/notes/notes.service.ts
@@ -9,12 +9,14 @@ export class NotesService {
       page = 1,
       limit = 10,
     }: PaginationParams = {},
+    user_id: schema.NoteInsertDTO['user_id'],
   ): Promise<PaginatedResponse<schema.NoteDTO>> {
     const offset = (page - 1) * limit;
 
     const [notes, total] = await Promise.all([
       db.select()
         .from(schema.notes)
+        .where(eq(schema.notes.user_id, user_id))
         .limit(limit)
         .offset(offset)
         .execute(),
@@ -23,17 +25,14 @@ export class NotesService {
         count: sql<number>`count(id)`,
       })
         .from(schema.notes)
+        .where(eq(schema.notes.user_id, user_id))
         .then((res) => res[0]?.count || 0),
     ]);
 
     const totalPages = Math.ceil(total / limit);
 
     return {
-      items: notes.map((note) => ({
-        id: note.id,
-        title: note.title,
-        content: note.content,
-      })),
+      items: notes,
       countItems: notes.length,
       total: +total,
       currentPage: page,
@@ -43,9 +42,9 @@ export class NotesService {
     };
   }
 
-  async getById(id: number) {
+  async getById(id: number, user_id: schema.NoteInsertDTO['user_id']) {
     const [note] = await db.select().from(schema.notes)
-      .where(eq(schema.notes.id, id))
+      .where(eq(schema.notes.id, id) && eq(schema.notes.user_id, user_id))
       .limit(1)
       .execute();
 
@@ -53,26 +52,35 @@ export class NotesService {
   }
 
 
-  async create(data: schema.NoteInsertDTO): Promise<schema.NoteDTO> {
+  async create(
+    data: Partial<Omit<schema.NoteInsertDTO, 'user_id'>>,
+    userId: schema.NoteInsertDTO['user_id'],
+  ): Promise<schema.NoteDTO> {
+    const value = { ...data, user_id: userId } as schema.NoteInsertDTO;
+
     const [note] = await db.insert(schema.notes)
-      .values(data)
+      .values(value)
       .returning();
 
     return note;
   }
 
-  async update(id: number, data: Partial<schema.NoteInsertDTO>): Promise<schema.NoteDTO | undefined> {
+  async update(
+    id: number,
+    data: Partial<Omit<schema.NoteInsertDTO, 'user_id'>>,
+    user_id: schema.NoteInsertDTO['user_id'],
+  ): Promise<schema.NoteDTO | undefined> {
     const [updatedNote] = await db.update(schema.notes)
       .set(data)
-      .where(eq(schema.notes.id, id))
+      .where(eq(schema.notes.id, id) && eq(schema.notes.user_id, user_id))
       .returning();
 
     return updatedNote;
   }
 
-  async delete(id: number): Promise<schema.NoteDTO | undefined> {
+  async delete(id: number, user_id: schema.NoteInsertDTO['user_id']): Promise<schema.NoteDTO | undefined> {
     const [deletedNote] = await db.delete(schema.notes)
-      .where(eq(schema.notes.id, id))
+      .where(eq(schema.notes.id, id) && eq(schema.notes.user_id, user_id))
       .returning();
 
     return deletedNote;
